pyproject.toml
[project]
name = "clue"
version = "0.1.0"
description = "Interactive Clue game with CrewAI"
requires-python = ">=3.12"
dependencies = [
    "crewai",
    "fastapi",
    "uvicorn",
    "python-dotenv",
    "pydantic",
    "jinja2",
    "python-multipart",
    "websockets"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/clue"]

[tool.uv]
dev-dependencies = [
    "pytest",
    "httpx"
]
